requirements.txt
pandas>=2.0.3
requests>=2.31.0
python-dotenv>=1.0.0
openpyxl>=3.1.2
python-slugify>=8.0.1